+++ tests/helpers/loading-from-directories/sample.type.ts
@@ -1,0 +1,6 @@
+import { Field, GraphQLObjectType } from "../../../src";
+
+@GraphQLObjectType()
+export class SampleObject {
+  @Field() sampleField: string;
+}